Moved validator functions to a new validators.py file.
